pyproject.toml
[project]
name = "campaign-manager"
version = "0.1.0"
description = "Marketing campaign manager with auto-batching, retry logic, and n8n integration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "gspread>=6.1.0",
    "google-auth>=2.29.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
]

[project.scripts]
campaign-manager = "app.main:run"
